Make NumberField internally uncontrolled and remove validation
NumberField currently handles internal validation of min max errors
Using defaultValue with the validation is prone to errors and edge cases
Additional work was needed to allow names and defaultValue because of prop restrictions
This commit lifts the TextField prop restrictions, omitting only SX and Type (type is restricted due to the use of type=number
All validation code has been removed, to be migrated to zod for use at the app level
All stories and tests related to the validations have been removed
The component, tests and stories have been moved to typescript
