+++ src/NumberField/NumberField.stories.tsx
@@ -1,30 +1,40 @@
+import { Meta, StoryFn, StoryObj } from "@storybook/react";
+import React, { ChangeEvent } from "react";
 import NumberField from "./NumberField";
-import React from "react";
+import { NumberFieldProps } from "./NumberField.types";
 import { action } from "@storybook/addon-actions";
+
+type Story = StoryObj<typeof NumberField>;
 
 export default {
   component: NumberField,
   title: "Text/NumberField"
-};
+} satisfies Meta<typeof NumberField>;
 
-const Template = args => {
+const Template: StoryFn<NumberFieldProps> = args => {
   const [value, setValue] = React.useState(args.value);
+
   React.useEffect(() => {
     setValue(args.value);
   }, [args.value]);
-  const onChange = event => {
+
+  const onChange = (
+    event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => {
     setValue(value);
     action("onChange")(event, event.target.value);
   };
-  const onBlur = event => {
+
+  const onBlur = () => {
     action("onBlur");
   };
+
   return (
     <NumberField {...args} onBlur={onBlur} onChange={onChange} value={value} />
   );
 };
 
-export const Default = {
+export const Default: Story = {
   args: {
     disabled: false,
     error: false,
@@ -32,7 +42,6 @@ export const Default = {
     label: "Enter a Number",
     margin: "normal",
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
     stepper: true,
     variant: "outlined"
@@ -41,7 +50,7 @@ export const Default = {
   render: Template
 };
 
-export const Uncontrolled = {
+export const Uncontrolled: Story = {
   args: {
     defaultValue: 12,
     disabled: false,
@@ -51,7 +60,6 @@ export const Uncontrolled = {
     margin: "normal",
     onBlur: action("onBlur"),
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
     stepper: true,
     variant: "outlined"
@@ -59,7 +67,7 @@ export const Uncontrolled = {
   render: NumberField
 };
 
-export const StartAdornment = {
+export const StartAdornment: Story = {
   args: {
     disabled: false,
     error: false,
@@ -67,7 +75,6 @@ export const StartAdornment = {
     label: "Enter a Number",
     margin: "normal",
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
     startAdornment: "$",
     stepper: true,
@@ -77,7 +84,7 @@ export const StartAdornment = {
   render: Template
 };
 
-export const EndAdornment = {
+export const EndAdornment: Story = {
   args: {
     disabled: false,
     endAdornment: "px",
@@ -86,7 +93,6 @@ export const EndAdornment = {
     label: "Enter a Number",
     margin: "normal",
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
     stepper: true,
     variant: "outlined"
@@ -95,7 +101,7 @@ export const EndAdornment = {
   render: Template
 };
 
-export const SmallAndDense = {
+export const SmallAndDense: Story = {
   args: {
     disabled: false,
     error: false,
@@ -103,7 +109,6 @@ export const SmallAndDense = {
     label: "Enter a Number",
     margin: "dense",
     required: false,
-    showMinMaxErrorMessage: true,
     size: "small",
     stepper: false,
     variant: "outlined"
@@ -112,7 +117,7 @@ export const SmallAndDense = {
   render: Template
 };
 
-export const NoStepper = {
+export const NoStepper: Story = {
   args: {
     disabled: false,
     error: false,
@@ -120,7 +125,6 @@ export const NoStepper = {
     label: "Enter a Number",
     margin: "normal",
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
     stepper: false,
     variant: "outlined"
@@ -129,55 +133,32 @@ export const NoStepper = {
   render: Template
 };
 
-export const CustomMinMaxAndStep = {
+export const CustomStep: Story = {
   args: {
     disabled: false,
     error: false,
+    inputProps: {
+      step: 0.1
+    },
     label: "Enter a Number",
     margin: "normal",
-    max: 1,
-    min: 0,
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
-    step: 0.1,
     variant: "outlined"
   },
 
   render: Template
 };
 
-export const InitialError = {
+export const InitialError: Story = {
   args: {
     disabled: false,
-    error: false,
+    error: true,
+    helperText: "Oh no you've made a mistake!",
     label: "Enter a Number",
     margin: "normal",
-    max: 1,
-    min: 0,
     required: false,
-    showMinMaxErrorMessage: true,
     size: "medium",
-    step: 0.1,
-    value: 2,
-    variant: "outlined"
-  },
-
-  render: Template
-};
-
-export const NoMinMaxErrorMessage = {
-  args: {
-    disabled: false,
-    error: false,
-    label: "Enter a Number",
-    margin: "normal",
-    max: 1,
-    min: 0,
-    required: false,
-    showMinMaxErrorMessage: false,
-    size: "medium",
-    step: 0.1,
     value: 2,
     variant: "outlined"
   },
